refactor(slides): clean/minimal by default + framework boundary
Two changes:

1. Most slides now default to a clean background + minimal chrome. Flipped the
   flag defaults (bg=clean, chrome=minimal); the cover keeps the full beam
   treatment via explicit data-bg="beam" data-chrome="full".

2. Gave the reusable slide engine a clear boundary: renamed src/systems ->
   src/framework and lifted the two engine pieces still embedded in main.js —
   the 16:9 stage fit and the magic-move FLIP engine — into framework/stage.js
   and framework/magic-move.js. main.js now imports them and drives everything
   through the single deckApi controller + deck:change event. Added
   framework/README.md documenting what is framework vs deck content (and the
   remaining navigation seam to lift next).
