+++ slides/index.html
@@ -41,7 +41,7 @@
   <!-- ====================================================
        1 — Cover
        ==================================================== -->
-  <section class="slide cover">
+  <section class="slide cover" data-bg="beam" data-chrome="full">
     <div class="cover__top">
       <div class="cover__title">
         <div class="cover__title-line">
@@ -788,7 +788,7 @@
   <!-- ====================================================
        27 — Thank you
        ==================================================== -->
-  <section class="slide thankyou" data-bg="clean" data-chrome="minimal">
+  <section class="slide thankyou">
     <span class="eyebrow">Fin.</span>
     <h1><span class="brand-text">Thank</span> you.</h1>
     <p class="sig">
