+++ slides/src/framework/icons.js
